% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_swiid.R
\name{download_swiid}
\alias{download_swiid}
\title{Download SWIID Dataset from Dataverse}
\usage{
download_swiid(dest_dir = "data", overwrite = FALSE)
}
\arguments{
\item{dest_dir}{Destination directory, default is "data"}

\item{overwrite}{Whether to overwrite existing file}
}
\value{
Path to downloaded .rda file
}
\description{
Download SWIID Dataset from Dataverse
}
\examples{
\dontrun{
# First set API key
set_api_key("your-api-key-here")

# Then download
download_swiid()
download_swiid(dest_dir = "my_data")
}
}
